+++ src/collision_constraint.hpp
@@ -17,23 +17,22 @@ struct CollisionConstraint {
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const = 0;
+        const double dhat) const = 0;
 
     virtual Eigen::VectorX12d compute_potential_gradient(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const = 0;
+        const double dhat) const = 0;
 
     virtual Eigen::MatrixXX12d compute_potential_hessian(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
         const double dhat,
-        const double dmin,
         const bool project_to_psd) const = 0;
+
+    double minimum_distance = 0;
 };
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -52,22 +51,19 @@ struct VertexVertexConstraint : VertexVertexCandidate, CollisionConstraint {
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::VectorX12d compute_potential_gradient(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::MatrixXX12d compute_potential_hessian(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
         const double dhat,
-        const double dmin,
         const bool project_to_psd) const override;
 
     long multiplicity = 1;
@@ -89,22 +85,19 @@ struct EdgeVertexConstraint : EdgeVertexCandidate, CollisionConstraint {
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::VectorX12d compute_potential_gradient(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::MatrixXX12d compute_potential_hessian(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
         const double dhat,
-        const double dmin,
         const bool project_to_psd) const override;
 
     long multiplicity = 1;
@@ -127,22 +120,19 @@ struct EdgeEdgeConstraint : EdgeEdgeCandidate, CollisionConstraint {
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::VectorX12d compute_potential_gradient(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::MatrixXX12d compute_potential_hessian(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
         const double dhat,
-        const double dmin,
         const bool project_to_psd) const override;
 
     double eps_x;
@@ -165,22 +155,19 @@ struct FaceVertexConstraint : FaceVertexCandidate, CollisionConstraint {
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::VectorX12d compute_potential_gradient(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
-        const double dhat,
-        const double dmin) const override;
+        const double dhat) const override;
 
     Eigen::MatrixXX12d compute_potential_hessian(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
         const Eigen::MatrixXi& F,
         const double dhat,
-        const double dmin,
         const bool project_to_psd) const override;
 };