+++ src/friction/friction.hpp
@@ -17,8 +17,7 @@ void construct_friction_constraint_set(
     double dhat,
     double barrier_stiffness,
     double mu,
-    FrictionConstraints& friction_constraint_set,
-    double dmin = 0);
+    FrictionConstraints& friction_constraint_set);
 
 /// @brief Compute the friction potential between to positions.
 ///